fix(ci): unblock v0.4.0 release pipeline
Three CI gates broke on the v0.4.0 push:

- clippy `unnecessary_cast`: GpuRef::raw_device_ptr returned
  `ptr as u64` but cudarc's device_ptr already returns u64.
- clippy `extra_unused_type_parameters`: copy_from_numpy_complex_typed
  / copy_to_numpy_complex_with had a leftover `S` generic from an
  earlier refactor; dropped.
- clippy `doc_lazy_continuation`: graph.rs module doc started a
  paragraph with a leading `+` which the markdown parser treated as
  a list bullet; reflowed to plain prose.

Also: maturin >=1.5 requires VIRTUAL_ENV; the python CI matrix now
creates a .venv before invoking `maturin develop`.

+++ crates/atomr-accel-cuda/src/gpu_ref.rs
@@ -162,7 +162,7 @@ impl<T> GpuRef<T> {
         // returned `u64` immediately on an FFI call. The underlying
         // CudaSlice<T> remains alive via the strong Arc held by
         // `self.inner.slice` for as long as this `GpuRef` lives.
-        Ok(ptr as u64)
+        Ok(ptr)
     }
 }
